volumen-ocr/requirements.txt
rapidocr-onnxruntime==1.4.4
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
opencv-python-headless==4.10.0.84